*** empty log message ***

+++ scripts/mergeAndRegister_online.py
@@ -4,26 +4,31 @@ import os
 import time
 import sys
 
+#DIR = '/afs/cern.ch/user/h/hkseo/public/DQM/dropbox'
 DIR = '/nfshome0/hkseo'
 if not os.path.exists(DIR +'/timeTag'):
     os.system('touch -t 01010000 '+ DIR +'/timeTag')
 
 ####### ENDLESS LOOP WITH SLEEP
 while 1:
-    
-    count = 0
-    for NEW in os.popen('find '+ DIR +'/ -type f -name "DQM_*_R?????????.root" -newer '+ DIR +'/timeTag'):
-        NEW = NEW.strip()
+
+    #### check new files
+    NEW_ = os.popen('find '+ DIR +'/ -type f -name "DQM_*_R?????????.root" -newer '+ DIR +'/timeTag').read().split()
+    if len(NEW_)==0:
+        #print 'waiting for new files...'
+        time.sleep(300)
+        continue
+
+    #### loop for new files
+    for NEW in NEW_:
         print NEW
-        count += 1
-        if count == 1:
-            os.system('touch '+ DIR +'/timeTag')
+        if NEW==NEW_[0]: os.system('touch '+ DIR +'/timeTag')
 
         #### extract run number of new file
         run = NEW[-14:-5]
 
         #### check if merged file exist & the size of the last merged file
-        FILE_EXIST = int(os.popen('ls '+ DIR + '/DQM_V* | wc').read().split()[0])
+        FILE_EXIST = len(os.popen('ls '+ DIR + '/DQM_V*').read().split())
         SIZE_LIMIT = 10000000
 
         if FILE_EXIST > 0:
@@ -59,6 +64,7 @@ while 1:
             ### merge new file
             time.sleep(1)
             os.system('hadd '+NEW_MERGED+' '+NEW+ ' '+FILE)
+            #os.system('touch '+NEW_MERGED)
             os.system('rm '+FILE)
         else:
             irun = run
@@ -66,16 +72,14 @@ while 1:
             NEW_MERGED = DIR + '/DQM_V001_R'+irun+'-R'+frun+'.root'
             os.system('cp '+NEW+' '+ NEW_MERGED)
 
-        ### File registration
-        TMP_DB = DIR + '/tmp/dqm.db'
-        os.system('set -e')
-        os.system('rm -f '+ TMP_DB)
-        if os.path.exists('/home/dqm/dqm.db'):
-            os.system('cp -p /home/dqm/dqm.db '+ TMP_DB)
-        os.system('visDQMRegisterFile '+ TMP_DB +' "Data taking" "CMS" '+ NEW_MERGED)
-        #os.system('mv '+ TMP_DB + ' /home/dqm/dqm-new.db')
-        #os.system('rm -f /home/dqm/dqm.db')
-        #os.system('mv /home/dqm/dqm-new.db /home/dqm/dqm.db')
-
-        time.sleep(300)
-        #time.sleep(1)
+        if NEW==NEW_[len(NEW_)-1]:
+            # File registration
+            TMP_DB = DIR + '/tmp/dqm.db'
+            os.system('set -e')
+            os.system('rm -f '+ TMP_DB)
+            if os.path.exists('/home/dqm/dqm.db'):
+                os.system('cp -p /home/dqm/dqm.db '+ TMP_DB)
+            os.system('visDQMRegisterFile '+ TMP_DB +' "Data taking" "CMS" '+ NEW_MERGED)
+            #os.system('mv '+ TMP_DB + ' /home/dqm/dqm-new.db')
+            #os.system('rm -f /home/dqm/dqm.db')
+            #os.system('mv /home/dqm/dqm-new.db /home/dqm/dqm.db')
